+++ CHANGELOG.md
@@ -1,3 +1,9 @@
 # EasyVariable Changelog
 
-## [Initial Version] - {PR_MERGE_DATE}
+## [Initial Version] - 2025-02-11
+- Support Youdao dictionary
+- Support Google translate
+- Support OpenAI API
+- Support Tencent translate
+- Support GLM (ZhipuAI)
+- Support Deepseek-v3